+++ Changelog.md
@@ -10,6 +10,7 @@
   - Apply the "-2" modifier to any Test Check while an Actor is "waiting".
 - Move some hard-coded numbers to their CONST equivalent references.
 - Newly-created Threats will have their prototype tokens NOT linked to the Actor.
+- Remove Initiative options from the Combat Tracker context menus (combat and combatant).
 
 ## v13.7.0 - Glory, Defeat and Intimidation